refactor: expose flood wait helper
