+++ telegram/downloader/verifier.go
@@ -9,8 +9,8 @@ import (
 	"golang.org/x/xerrors"
 
 	"github.com/gotd/td/internal/crypto"
-	"github.com/gotd/td/telegram/internal/helpers"
 	"github.com/gotd/td/tg"
+	"github.com/gotd/td/tgerr"
 )
 
 // ErrHashMismatch means that download hash verification was failed.
@@ -87,7 +87,7 @@ func (v *verifier) next(ctx context.Context) (tg.FileHash, bool, error) {
 
 	for {
 		hashes, err := v.client.Hashes(ctx, v.offset)
-		if flood, err := helpers.FloodWait(ctx, err); err != nil {
+		if flood, err := tgerr.FloodWait(ctx, err); err != nil {
 			if flood {
 				continue
 			}
