+++ telegram/flood_wait.go
@@ -1,15 +1,15 @@
 package telegram
 
 import (
-	"github.com/gotd/td/telegram/internal/helpers"
+	"github.com/gotd/td/tgerr"
 )
 
 // ErrFloodWait is error type of "FLOOD_WAIT" error.
-const ErrFloodWait = helpers.ErrFloodWait
+const ErrFloodWait = tgerr.ErrFloodWait
 
 // AsFloodWait returns wait duration and true boolean if err is
 // the "FLOOD_WAIT" error.
 //
 // Client should wait for that duration before issuing new requests with
 // same method.
-var AsFloodWait = helpers.AsFloodWait // nolint:gochecknoglobals
+var AsFloodWait = tgerr.AsFloodWait // nolint:gochecknoglobals
