+++ telegram/internal/e2etest/terentyev.go
@@ -8,9 +8,9 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/xerrors"
 
-	"github.com/gotd/td/telegram/internal/helpers"
 	"github.com/gotd/td/telegram/message"
 	"github.com/gotd/td/tg"
+	"github.com/gotd/td/tgerr"
 )
 
 // User is a simple user bot.
@@ -78,7 +78,7 @@ func (u User) Run(ctx context.Context) error {
 			time.Sleep(2 * time.Second)
 
 			_, err = sender.To(peer).Text(ctx, line)
-			if flood, err := helpers.FloodWait(ctx, err); err != nil {
+			if flood, err := tgerr.FloodWait(ctx, err); err != nil {
 				if flood {
 					continue
 				}
